app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
psycopg2-binary>=2.9.9

# Database Migrations (Flyway-like)
yoyo-migrations>=9.0.0

# LangChain Ecosystem (versiones compatibles verificadas)
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.40  # Requerido por langchain-openai 0.1.25
langchain-openai==0.1.25
langgraph==0.1.19
# langmem==0.0.28  # TODO: Añadir cuando se implemente memoria (requiere langchain-anthropic>=0.3.3)

# LangChain Providers
langchain-anthropic==0.1.23  # Claude (future)
langchain-cohere==0.1.9      # Cohere (future)

# Embeddings
sentence-transformers>=3.0.0
scikit-learn==1.3.2
huggingface-hub>=0.20.0
# transformers>=5.5 enforces torch>=2.6.0 due to CVE-2025-32434
torch>=2.6.0

# OCR & PDF Processing
pdf2image==1.16.3
Pillow==10.1.0
opencv-python==4.6.0.66
pytesseract>=0.3.10
PyMuPDF>=1.23.0  # For PDF analysis and text detection

# Docker Management
docker>=7.0.0  # For restarting external Tika container

# Qdrant
qdrant-client==1.15.1

# LLM
ollama==0.1.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0

# Backup & Scheduling
apscheduler>=3.10.0
requests>=2.31.0

# Utils
numpy==1.24.3
pyyaml==6.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0  # Coverage reports
pytest-mock>=3.11.0  # Mocking helpers
